Phase C: replace Win32 API with Cocoa and build on macOS
- Port MessageBox/FileDialog/Locale to Objective-C++ (.mm) using NSAlert,
  NSOpenPanel/NSSavePanel, and NSLocale
- Drop wide-string code paths; route all I/O through UTF-8 std::string
- Replace ma_sound_init_from_file_w with UTF-8 overload
- Normalize res\\ backslash paths to res/
- Rename Math.{h,cpp} to MathUtils.{h,cpp} to avoid case-insensitive
  filesystem collision with <math.h>
- Stub SAL annotations for DirectXMath under non-Windows builds
- Enable ImGui ConfigMacOSXBehaviors and remap Ctrl to Super at
  input-check time so existing shortcut config stays cross-platform
- Drop unsupported std::execution::par/seq policies (Apple Clang libc++
  has no TBB backend)
- Bundle .mm sources and copy res/ into Contents/Resources/res/ via CMake

Empty-chart launch verified on macOS; .app bundle runs past init.

+++ MikuMikuWorld/Audio/Waveform.h
@@ -4,7 +4,7 @@
 */
 
 #pragma once
-#include "../Math.h"
+#include "../MathUtils.h"
 #include "AudioManager.h"
 #include <stdint.h>
 #include <vector>
@@ -45,7 +45,7 @@ namespace Audio
 			const float sampleIndexAsFloat = static_cast<float>(seconds * samplesPerSecond);
 			const float sampleIndexFraction = sampleIndexAsFloat - static_cast<int32_t>(sampleIndexAsFloat);
 
-			const size_t sampleIndexLo = std::max(0ull, static_cast<size_t>(sampleIndexAsFloat));
+			const size_t sampleIndexLo = std::max(static_cast<size_t>(0), static_cast<size_t>(sampleIndexAsFloat));
 			const size_t sampleIndexHi = sampleIndexLo + 1;
 
 			if (sampleIndexLo >= absoluteSamples.size())
